# Optimize Ukraine observed holidays