Performance tests:
```
country = "UA"
START_YEAR = 1917
END_YEAR = 2099

def calc_hol1():
    for year in range(START_YEAR, END_YEAR + 1):
        hols = holidays.country_holidays(country, years=year)

def calc_hol2():
    hols = holidays.country_holidays(country, years=range(START_YEAR, END_YEAR + 1))

t1 = timeit.timeit(calc_hol1, number=100)
t2 = timeit.timeit(calc_hol2, number=100)
```
| Test | Current beta |   PR  |
|:----:|:------------:|:-----:|
|  t1  |        2.290 | 2.219 |
|  t2  |       19.376 | 1.925 |